requirements.txt
Django>=4.2
psycopg2-binary
Pillow>=10.0
djangorestframework
django-cors-headers